+++ client/src/util/constants.js
@@ -1,2 +1,6 @@
 export const iconWidth = 25;
 export const marginSize = 7;
+export const domesticIslandDistance = 100;
+export const islandInterpackageDistance = 600;
+export const manyBodyStrength = -1000; // negative is repulsion
+export const simIterations = 500;
